+++ react-client/src/helpers/validation.js
@@ -80,3 +80,8 @@ export const validateInput = input => {
   });
   return valid;
 };
+
+export const validateEmail = input => {
+  let regex = /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/;
+  return regex.test(input);
+}
