+++ react-client/src/index.css
@@ -1,5 +1,6 @@
 @import url("https://fonts.googleapis.com/css?family=Permanent+Marker&display=swap");
 @import url("https://fonts.googleapis.com/css?family=Oxygen&display=swap");
+@import url("https://fonts.googleapis.com/css?family=Quicksand&display=swap");
 
 body {
   margin: 0;
